+++ source/frontend/commands/bounty/verify-turn-in.js
@@ -1,7 +1,6 @@
 const { userMention, bold, MessageFlags, ActionRowBuilder, StringSelectMenuBuilder, UserSelectMenuBuilder } = require("discord.js");
-const { SelectMenuLimits } = require("@sapphire/discord.js-utilities");
 const { SubcommandWrapper } = require("../../classes");
-const { listifyEN, commandMention, updatePosting, congratulationBuilder, truncateTextToLength, disabledSelectRow } = require("../../shared");
+const { listifyEN, commandMention, updatePosting, congratulationBuilder, disabledSelectRow, bountiesToSelectOptions } = require("../../shared");
 const { timeConversion } = require("../../../shared");
 const { SKIP_INTERACTION_HANDLING, SAFE_DELIMITER } = require("../../../constants");
 
@@ -13,20 +12,13 @@ module.exports = new SubcommandWrapper("verify-turn-in", "Verify up to 5 bounty 
 			return;
 		}
 
-		const options = openBounties.map(bounty => {
-			const option = { label: bounty.title, value: bounty.id };
-			if (bounty.description) {
-				option.description = truncateTextToLength(bounty.description, SelectMenuLimits.MaximumLengthOfDescriptionOfOption);
-			}
-			return option;
-		});
 		interaction.reply({
 			content: `Select one of your bounties and some bounty hunters who have turned it in. XP and drops will be distributed to those hunters when you use ${commandMention("bounty complete")}.`,
 			components: [
 				new ActionRowBuilder().addComponents(
 					new StringSelectMenuBuilder().setCustomId(`${SKIP_INTERACTION_HANDLING}${SAFE_DELIMITER}bounty`)
 						.setPlaceholder("Select a bounty...")
-						.addOptions(options)
+						.addOptions(bountiesToSelectOptions(openBounties))
 				),
 				new ActionRowBuilder().addComponents(
 					new UserSelectMenuBuilder().setCustomId(SKIP_INTERACTION_HANDLING)
@@ -39,12 +31,12 @@ module.exports = new SubcommandWrapper("verify-turn-in", "Verify up to 5 bounty 
 		}).then(response => response.resource.message).then(message => {
 			const collector = message.createMessageComponentCollector({ time: timeConversion(2, "m", "ms") });
 			let bounty;
-			collector.on("collect", async collectedInteration => {
-				const [_, stepId] = collectedInteration.customId.split(SAFE_DELIMITER);
+			collector.on("collect", async collectedInteraction => {
+				const [_, stepId] = collectedInteraction.customId.split(SAFE_DELIMITER);
 				switch (stepId) {
 					case "bounty":
-						bounty = openBounties.find(bounty => bounty.id === collectedInteration.values[0]);
-						collectedInteration.update({
+						bounty = openBounties.find(bounty => bounty.id === collectedInteraction.values[0]);
+						collectedInteraction.update({
 							components: [
 								disabledSelectRow(bounty.title),
 								new ActionRowBuilder().addComponents(
@@ -57,8 +49,8 @@ module.exports = new SubcommandWrapper("verify-turn-in", "Verify up to 5 bounty 
 						break;
 					case "hunters":
 						try {
-							let { bounty: returnedBounty, allCompleters, company, validatedCompleterIds, bannedIds } = await logicLayer.bounties.addCompleters(bounty, interaction.guild, [...collectedInteration.members.values()], runMode);
-							const post = await updatePosting(collectedInteration.guild, company, returnedBounty, poster.getLevel(company.xpCoefficient), allCompleters);
+							let { bounty: returnedBounty, allCompleters, company, validatedCompleterIds, bannedIds } = await logicLayer.bounties.addCompleters(bounty, interaction.guild, [...collectedInteraction.members.values()], runMode);
+							const post = await updatePosting(collectedInteraction.guild, company, returnedBounty, poster.getLevel(company.xpCoefficient), allCompleters);
 							const sentences = [];
 							if (validatedCompleterIds.length > 0) {
 								sentences.push(`The following bounty hunters' turn-ins of ${bold(returnedBounty.title)} have been recorded: ${listifyEN(validatedCompleterIds.map(id => userMention(id)))}`);
@@ -69,7 +61,7 @@ module.exports = new SubcommandWrapper("verify-turn-in", "Verify up to 5 bounty 
 							if (bannedIds.length > 0) {
 								sentences.push(`The following users were skipped due to currently being banned from using BountyBot: ${listifyEN(bannedIds.map(id => userMention(id)))}`);
 							}
-							collectedInteration.update({
+							collectedInteraction.update({
 								content: sentences.join("\n\n"),
 								components: []
 							});
@@ -77,7 +69,7 @@ module.exports = new SubcommandWrapper("verify-turn-in", "Verify up to 5 bounty 
 							if (typeof e !== 'string') {
 								console.error(e);
 							} else {
-								collectedInteration.update({ content: e, components: [] });
+								collectedInteraction.update({ content: e, components: [] });
 							}
 							return;
 						}