+++ ChangeLog.md
@@ -3,9 +3,9 @@
 ### User Options in Slash Commands
 Slash commands originally accepted users as a string (instead of Discord's user filtering) to allow users to mention as many users as they wanted. However, this doesn't work on mobile, so those slash commands have been remade to use Discord's user filtering or message selects.
 - `/toast` now requires 1 toastee and up to 4 optional ones. The `message` option has been changed to the first option to group the toastee options.
-- `/bounty add-completers` has been renamed to `/bounty verify-turn-in`. It receives both its bounty and hunter inputs from message selects.
+- `/bounty add-completers` has been renamed to `/bounty verify-turn-in` and now uses message selects.
+- `/bounty remove-completers` has been renamed to `/bounty revoke-turn-in` and now uses message selects.
 - `/bounty complete` now accepts up to 5 optional hunters.
-- `/bounty remove-completers` has been renamed to `/bounty revoke-turn-in` and now requires 1 hunter and up to 4 optional ones.
 - `/evergreen complete` now requires 1 hunter and up to 4 optional ones. Duplicate hunters are now accepted in case the hunter has multiple turn-ins to be awarded for.
 ### Premium
 - Added the ability to customize the Goal Completion embed's thumbnail with `/config-premium goal-completion-thumbnail-url`